Change Safari version tests from 8 to 7
Safari 8 will not pass tests in Sauce Labs because it ends in a "DISCONNECT" status and exits make with a 1.  Removing blocks in the contact page Django/Swig template seems to "fix" the problem.  Since this seems to be a webdriver bug or something rather than my tests being broken I am settling on running JS unit tests in Safari 7 instead of 8