Upgrade WebdriverIO to 3.0, upgrade PhantomJS to 2.0 to support it
It doesn't seem to work with Phantom 1.9.x, see https://github.com/ariya/phantomjs/issues/11421

The phantomjs package.json entry is using the forked phantom2 package, at the commit right before the package's name was changed from "phantomjs" to "phantomjs2"—allowing phantomjs2 to satisfy the karma-phantomjs-launcher peer dependency requirement.  Note that this does not work in Linux.  When the official phantomjs npm package is upgraded to 2.0 it would be a good idea to switch back to it.